% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipoints.R
\name{mk_sfc_MULTIPOINT}
\alias{mk_sfc_MULTIPOINT}
\title{Make vector of multi points}
\usage{
mk_sfc_MULTIPOINT(x, objectindex = NULL, ..., crs = NA_character_)
}
\arguments{
\item{x}{coordinates of multi points}

\item{objectindex}{starting index of each object (if more than one)}

\item{...}{reserved}

\item{crs}{the projection, EPSG code (integer) or proj4string}
}
\description{
Make vector of multi points
}
