% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{mk_sfc_POINT}
\alias{mk_sfc_POINT}
\title{Make vector of points}
\usage{
mk_sfc_POINT(x, crs = NA, ...)
}
\arguments{
\item{x}{coordinates of points}

\item{crs}{coordinate system in EPSG (integer) or proj4string}

\item{...}{reserved}
}
\description{
Make vector of points
}
