% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{mk_sfc_POLYGON}
\alias{mk_sfc_POLYGON}
\title{Make vector of polygons}
\usage{
mk_sfc_POLYGON(xlist, crs = NA, ...)
}
\arguments{
\item{crs}{coordinate system in EPSG (integer) or proj4string}

\item{...}{reserved}

\item{x}{coordinates of polygons}
}
\description{
Make vector of polygons
}
