% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipoints.R
\name{multipoints_rcpp}
\alias{multipoints_rcpp}
\title{Make multi points}
\usage{
multipoints_rcpp(pts, objectindex = NULL, ..., gdim = "XY")
}
\arguments{
\item{pts}{coordinates input coordinates for points}

\item{objectindex}{starting index of each object (if more than one)}

\item{...}{reserved}

\item{gdim}{the geometric dimension in `sf::dimension` terms, default is XY}
}
\description{
Make multi points
}
