% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.R
\name{points_rcpp}
\alias{points_rcpp}
\title{Make points}
\usage{
points_rcpp(pts, gdim = "XY", ...)
}
\arguments{
\item{pts}{list of point vectors (assumed length 2)}

\item{gdim}{geometric dimension (in `sf::st_dimension` terms)}

\item{...}{reserved}
}
\description{
Make points
}
