% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygons.R
\name{polygons_rcpp}
\alias{polygons_rcpp}
\title{Make polygons}
\usage{
polygons_rcpp(xlist, gdim = "XY", ...)
}
\arguments{
\item{xlist}{list of polygon matrices}

\item{gdim}{geometric dimension (in `sf::st_dimension` terms)}

\item{...}{reserved}
}
\description{
Make polygons
}
