% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfcc-package.R
\docType{package}
\name{sfcc-package}
\alias{sfcc}
\alias{sfcc-package}
\title{sfcc: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Michael Sumner \email{mdsumner@gmail.com}

}
\keyword{internal}
